Implement the q-def shortcode and add processing for shortcode in YAML
This is the q-def shortcode for glossary pop-ups that we've used in Quire books before, but with a couple notable differences. First, it pulls the definition from the individual vocabulary markdown pages rather than a references.yml file. Second, there are pop-ups within definitions, which are store as YAML. Becaue shortcodes in YAML are not normally processed, this commit adds some find/replace logic to manually process the shortcodes in the YAML as it's output on the vocabularly pages.